src/main/webapp/app/entities/BulutAttendance/record-status/delete/record-status-delete-dialog.component.spec.ts
jest.mock('@ng-bootstrap/ng-bootstrap');

import { ComponentFixture, TestBed, inject, fakeAsync, tick } from '@angular/core/testing';
import { HttpResponse } from '@angular/common/http';
import { HttpClientTestingModule } from '@angular/common/http/testing';
import { of } from 'rxjs';
import { NgbActiveModal } from '@ng-bootstrap/ng-bootstrap';

import { RecordStatusService } from '../service/record-status.service';

import { RecordStatusDeleteDialogComponent } from './record-status-delete-dialog.component';

describe('RecordStatus Management Delete Component', () => {
  let comp: RecordStatusDeleteDialogComponent;
  let fixture: ComponentFixture<RecordStatusDeleteDialogComponent>;
  let service: RecordStatusService;
  let mockActiveModal: NgbActiveModal;

  beforeEach(() => {
    TestBed.configureTestingModule({
      imports: [HttpClientTestingModule, RecordStatusDeleteDialogComponent],
      providers: [NgbActiveModal],
    })
      .overrideTemplate(RecordStatusDeleteDialogComponent, '')
      .compileComponents();
    fixture = TestBed.createComponent(RecordStatusDeleteDialogComponent);
    comp = fixture.componentInstance;
    service = TestBed.inject(RecordStatusService);
    mockActiveModal = TestBed.inject(NgbActiveModal);
  });

  describe('confirmDelete', () => {
    it('Should call delete service on confirmDelete', inject(
      [],
      fakeAsync(() => {
        // GIVEN
        jest.spyOn(service, 'delete').mockReturnValue(of(new HttpResponse({ body: {} })));

        // WHEN
        comp.confirmDelete('9fec3727-3421-4967-b213-ba36557ca194');
        tick();

        // THEN
        expect(service.delete).toHaveBeenCalledWith('9fec3727-3421-4967-b213-ba36557ca194');
        expect(mockActiveModal.close).toHaveBeenCalledWith('deleted');
      }),
    ));

    it('Should not call delete service on clear', () => {
      // GIVEN
      jest.spyOn(service, 'delete');

      // WHEN
      comp.cancel();

      // THEN
      expect(service.delete).not.toHaveBeenCalled();
      expect(mockActiveModal.close).not.toHaveBeenCalled();
      expect(mockActiveModal.dismiss).toHaveBeenCalled();
    });
  });
});
